Make the following changes:
,

Find this element:

  <Button variant="default" size="default" aria-label="Link action">

with the text content:

    Link

Relevant files:

    - https://6000-firebase-studio-1748371280287.cluster-6vyo4gb53jczovun3dxslzjahs.cloudworkstations.dev/_next/static/chunks/src_aa4cf66b._.js
    - https://6000-firebase-studio-1748371280287.cluster-6vyo4gb53jczovun3dxslzjahs.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1748371280287.cluster-6vyo4gb53jczovun3dxslzjahs.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

change this button to human icon, as I will use this select participant in the meeting room. This button once clicked, will display a drop down window as participant options. The options should contains all roles in those pre defined meetings